1.Render在注册 render 时候会携带一个内置的 servlet。 2.Servlets 支持性发现严重问题，先递交版本在调优功能。

+++ src/main/java/net/hasor/web/invoker/InvokerWebApiBinder.java
@@ -340,9 +340,9 @@ public class InvokerWebApiBinder extends ApiBinderWrap implements WebApiBinder {
     }
     //
     // ------------------------------------------------------------------------------------------------------
-    protected void jeeServlet(int index, String pattern, UriPatternMatcher matcher, BindInfo<? extends HttpServlet> servletRegister, Map<String, String> initParams) {
+    protected void jeeServlet(long index, String pattern, UriPatternMatcher matcher, BindInfo<? extends HttpServlet> servletRegister, Map<String, String> initParams) {
         ServletDefinition define = new ServletDefinition(index, pattern, matcher, servletRegister, initParams);
-        bindType(ServletDefinition.class).uniqueName().toInstance(define);/*单列*/
+        bindType(ServletDefinition.class).uniqueName().toInstance(define);/*单例*/
     }
     @Override
     public ServletBindingBuilder jeeServlet(final String urlPattern, final String... morePatterns) {
